# PERF: reduce overhead in groupby _cython_operation