@jbrockmendel 2021-03-09
can you also report how the non-AM perf is affected?