@jorisvandenbossche 2021-03-09
(some failing tests related to (unsupported) operations on categorical column that I still need to address)